+++ src/main/java/com/apex/repository/ITGUserRepository.java
@@ -3,5 +3,5 @@ package com.apex.repository;
 import com.apex.entities.TGUser;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ITGUserRepository extends CrudRepository<TGUser, Integer> {
+public interface ITGUserRepository extends CrudRepository<TGUser, Long> {
 }
